--- components/Login.tsx
@@ -1,19 +1,0 @@
-import { useSession, signIn, signOut } from "next-auth/react"
-
-export default function Login() {
-  const { data: session } = useSession()
-  if (session?.user) {
-    return (
-      <>
-        Signed in as {session.user.email} <br />
-        <button onClick={() => signOut()}>Sign out</button>
-      </>
-    )
-  }
-  return (
-    <>
-      Not signed in <br />
-      <button onClick={() => signIn()}>Sign in</button>
-    </>
-  )
-}
